requirements.txt
# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0

# Object Detection
ultralytics>=8.0.0

# Computer Vision
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
Pillow>=10.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Data Augmentation
albumentations>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress Bars
tqdm>=4.65.0

# Configuration
PyYAML>=6.0

# Experiment Tracking
tensorboard>=2.13.0

# Utilities
scikit-learn>=1.3.0
scipy>=1.11.0

# Annotation Tools (optional)
# labelImg>=1.8.6

# API Development (optional for deployment)
# fastapi>=0.100.0
# uvicorn>=0.23.0

# Image handling
imageio>=2.31.0

# Metrics
pycocotools>=2.0.6

# GPU acceleration (optional)
# nvidia-ml-py3>=7.352.0
